pyproject.toml
[tool.poetry]
name = "llmops-with-haystack"
version = "0.1.0"
description = ""
authors = ["Laura Gutierrez Funderburk <lgutierrwr@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
farm-haystack = {extras = ["faiss", "inference", "preprocessing", "weaviate"], version = "1.21.1"}
chainlit = "^0.7.0"
openai = "^0.28.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.2"
python-dotenv = "^1.0.0"
datasets = "^2.14.5"
kaggle = "^1.5.16"
flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
